requirements.txt
opencv-python==4.6.0.66
numpy==1.23.5
Pillow==9.3.0